Move `RenderEditorHistory` function to `editor_history.cpp`
